backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-cors==5.0.0
psycopg2-binary==2.9.10
werkzeug==3.1.3
newspaper3k==0.2.8
beautifulsoup4==4.13.3
transformers==4.47.0
torch==2.5.1
lime==0.2.0.1
python-dotenv==1.0.1
gunicorn==23.0.0
